Rename TapisAuthenticator to Authenticator and update old refs
