+++ src/packages/tapis/TapisController.py
@@ -4,7 +4,7 @@ from tapipy.tapis import Tapis
 
 from conf import settings
 from core.Controller import Controller
-from packages.tapis.TapisAuthenticator import TapisAuthenticator as Auth
+from packages.tapis.Authenticator import Authenticator as Auth
 
 
 class TapisController(Controller):
