+++ src/packages/tapis/controllers/Auth.py
@@ -5,6 +5,7 @@ from utils.ConfigManager import ConfigManager
 from utils.Prompt import prompt
 import conf.settings as settings
 from packages.tapis.settings import ENVS
+from packages.core.controllers.Setup import Setup
 
 
 class Auth(Controller):
@@ -20,10 +21,10 @@ class Auth(Controller):
         provide the credentials for the defined authentication method.
         """
         # If the configs.ini specified in the settings does not exist,
-        # create it.
+        # call the init method on the setup controller.
         if not os.path.isfile(settings.CONFIG_FILE):
-            self.logger.log(f"Creating config file '{settings.CONFIG_FILE}'\n")
-            self.conf.create_config_file()
+            setup_controller = Setup()
+            setup_controller.init()
 
         # Create the credentials section if it doesn't exsit. (It's possible 
         # that the credentials section has been erased even though
